Added CreateApp method

+++ src/main/java/com/example/BLPS/Repositories/PlatformRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface PlatformRepository extends JpaRepository<Platform, Long> {
+public interface PlatformRepository extends JpaRepository<Platform, Integer> {
     Platform findPlatformByName(String name);
 }
